Refactor CSD adapter
The .convertCSDToContact method should really be in the adapter object.
